# stmtsummary: optimize performance