--- a/pkg/util/stmtsummary/statement_summary.go
+++ b/pkg/util/stmtsummary/statement_summary.go
@@ -318,7 +318,7 @@ func newStmtSummaryByDigestMap() *stmtSummaryByDigestMap {
 	return newSsMap
 }
 
-// AddStatement adds a statement to StmtSummaryByDigestMap.
+// AddStatementadds a statement to StmtSummaryByDigestMap.
 func (ssMap *stmtSummaryByDigestMap) AddStatement(sei *StmtExecInfo) {
 	// All times are counted in seconds.
 	now := time.Now().Unix()
@@ -345,40 +345,43 @@ func (ssMap *stmtSummaryByDigestMap) AddStatement(sei *StmtExecInfo) {
 	key.Init(sei.SchemaName, sei.Digest, sei.PrevSQLDigest, sei.PlanDigest, sei.ResourceGroupName)
 
 	var exist bool
-	// Enclose the block in a function to ensure the lock will always be released.
-	summary, beginTime := func() (*stmtSummaryByDigest, int64) {
-		ssMap.Lock()
-		defer ssMap.Unlock()
 
-		// Check again. Statements could be added before disabling the flag and after Clear().
-		if !ssMap.Enabled() {
-			return nil, 0
-		}
-		if sei.IsInternal && !ssMap.EnabledInternal() {
-			return nil, 0
-		}
+	// Using a global lock here instead of fine-grained locks because:
+	// 1. The critical sections are very short, making lock overhead potentially higher 
+	// than the protected code execution time.
+	// 2. Previous implementation with layered locks reveals significant contention and 
+	// poorer performance in benchmarks.
+	// A single coarse-grained lock reduces overall contention and may provide better
+	// performance in this specific case.
+	ssMap.Lock()
+	defer ssMap.Unlock()
 
-		if ssMap.beginTimeForCurInterval+intervalSeconds <= now {
-			// `beginTimeForCurInterval` is a multiple of intervalSeconds, so that when the interval is a multiple
-			// of 60 (or 600, 1800, 3600, etc), begin time shows 'XX:XX:00', not 'XX:XX:01'~'XX:XX:59'.
-			ssMap.beginTimeForCurInterval = now / intervalSeconds * intervalSeconds
-		}
+	// Check again. Statements could be added before disabling the flag and after Clear().
+	if !ssMap.Enabled() {
+		return
+	}
+	if sei.IsInternal && !ssMap.EnabledInternal() {
+		return
+	}
 
-		beginTime := ssMap.beginTimeForCurInterval
-		var value kvcache.Value
-		value, exist = ssMap.summaryMap.Get(key)
-		var summary *stmtSummaryByDigest
-		if !exist {
-			// Lazy initialize it to release ssMap.mutex ASAP.
-			summary = new(stmtSummaryByDigest)
-			ssMap.summaryMap.Put(key, summary)
-		} else {
-			summary = value.(*stmtSummaryByDigest)
-		}
-		summary.isInternal = summary.isInternal && sei.IsInternal
-		return summary, beginTime
-	}()
-	// Lock a single entry, not the whole cache.
+	if ssMap.beginTimeForCurInterval+intervalSeconds <= now {
+		// `beginTimeForCurInterval` is a multiple of intervalSeconds, so that when the interval is a multiple
+		// of 60 (or 600, 1800, 3600, etc), begin time shows 'XX:XX:00', not 'XX:XX:01'~'XX:XX:59'.
+		ssMap.beginTimeForCurInterval = now / intervalSeconds * intervalSeconds
+	}
+
+	beginTime := ssMap.beginTimeForCurInterval
+	var value kvcache.Value
+	value, exist = ssMap.summaryMap.Get(key)
+	var summary *stmtSummaryByDigest
+	if !exist {
+		// Lazy initialize it to release ssMap.mutex ASAP.
+		summary = new(stmtSummaryByDigest)
+		ssMap.summaryMap.Put(key, summary)
+	} else {
+		summary = value.(*stmtSummaryByDigest)
+	}
+	summary.isInternal = summary.isInternal && sei.IsInternal
 	if summary != nil {
 		summary.add(sei, beginTime, intervalSeconds, historySize)
 	}