<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #60305

Problem Summary:

### What changed and how does it work?

Use a more corase-grained locking to reduce mutex overhead in stmt summary.

vanilla
```
go test -bench=BenchmarkStmtSummaryAdd -benchtime=10s
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/pkg/util/stmtsummary/v2
cpu: AMD Ryzen 9 9900X 12-Core Processor            
BenchmarkStmtSummaryAddSingleWorkload-24            	80699414	       149.9 ns/op
BenchmarkStmtSummaryAddParallelSingleWorkload-24    	14539602	       757.8 ns/op
BenchmarkStmtSummaryAddParallelMultiWorkload-24     	19443807	       620.6 ns/op
PASS
ok  	github.com/pingcap/tidb/pkg/util/stmtsummary/v2	36.834s
```

This PR
```
go test -bench=BenchmarkStmtSummaryAdd -benchtime=10s
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/pkg/util/stmtsummary/v2
cpu: AMD Ryzen 9 9900X 12-Core Processor            
BenchmarkStmtSummaryAddSingleWorkload-24            	98894013	       122.5 ns/op
BenchmarkStmtSummaryAddParallelSingleWorkload-24    	48242661	       240.1 ns/op
BenchmarkStmtSummaryAddParallelMultiWorkload-24     	48644108	       246.5 ns/op
PASS
ok  	github.com/pingcap/tidb/pkg/util/stmtsummary/v2	36.356s
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test: existing tests
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```