@Defined2014 2025-04-07
> I extracted `affectedRows` from `stmtCtxMu` for performance, according to the benchmark. However, I'm open to reverting this change if it results in poor design or introduces unnecessary risks. cc @Defined2014

LGTM, seems the `TestIssue58600` test needs to modify, maybe we could move the failpoint to another function like `AddFoundRows()`